(REST) Remove rest module
