+++ sensormesh/thingspeak.py
@@ -3,8 +3,7 @@ from datetime import datetime
 import requests
 import dateutil.parser
 
-from .endpoints import DataApi
-from .rest import RestSource, RestTarget
+from .endpoints import DataTarget, DataSource, ApiMixin, DataApi
 from .exceptions import ConfigurationError
 
 
@@ -76,7 +75,7 @@ class ThingSpeakApi(DataApi):
             return None
 
 
-class ThingSpeakLogger(RestTarget):
+class ThingSpeakLogger(ApiMixin, DataTarget):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, api_cls=ThingSpeakApi, **kwargs)
 
@@ -92,8 +91,11 @@ class ThingSpeakLogger(RestTarget):
 
         return data_out
 
+    def _update(self, data):
+        self._api.post_update(data)
 
-class ThingSpeakSource(RestSource):
+
+class ThingSpeakSource(ApiMixin, DataSource):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, api_cls=ThingSpeakApi, **kwargs)
 
@@ -106,3 +108,6 @@ class ThingSpeakSource(RestSource):
             data['timestamp'] = ts.timestamp()
 
         return data
+
+    def _read(self):
+        return self._api.get_data()
